refactor(utils): 유틸 파일명 변경
