+++ src/utils/searchHistoryUtils.ts
